refactor(bridge): extract action diagnostics
